INTERLOK-1249 Initial Checkin INTERLOK-1249 #time 4h
The examples suggest you use PreBuiltTransportClient but that really screws
up the dependencies; made a clone of the PreBuilt that just uses Netty4Plugin
as the baseline plugin.
Other than resolving the API change; this is basically a cut/paste job
from the original elastic-search project.
Because ES5.0.0 doesn't seem to declare its lucene dependencies properly we
have had to manually declare them all for 6.2.0 (which matches the ES5.0.0
distr).